autov: add field definitions for pa1 and set weight to 1
PA1 requires the field definitions present in a "clean copy" (which contains
field definitions for PA2) to be flipped in X. We do that now in set_fields. We
also neglected to set the fields to 1, which they start at some strange values.
Do that at the same time no matter the array.